Pricing fix + start of allowance
int > double
start of allowance calculation

+++ controller/SalesController.java
@@ -9,7 +9,7 @@ public class SalesController {
         salesContainer = salesContainer.getInstance();
     }
     
-    public boolean createSale(int finalPrice, String date, String salesNumber){
+    public boolean createSale(double finalPrice, String date, String salesNumber){
         Sale sale = new Sale(finalPrice, date, salesNumber);
         salesContainer.addSale(sale);
         return true;
